PR: Fixed issues for World Space / Animated UIs
Passing position and rotation into Instantiate uses *world* position/rotation, so this only works when the UI is at origin / not animated. With an animated world space UI, the problem with passing those in is that the new UI elements end up being offset (position/rotation) from the main UI and either float in space or are behind the camera.

With ClickOffCatcher, position, rotation and scale had to be set correctly (to local zero/identity/one).