PR: Restored ClearFilters() because it is still used in the UI
The button _ModBrowser / MainContainer+Mask / ExplorerView / MainPanel / NoResults / ClearFilterButton didn't work. The reason is probably due to a refactoring where ClearFilters() was renamed to ClearAllFilters(), so the event link broke. I have (re-)added ClearFilters() and marked it as obsolete. Alternatively, the prefab could be changed to use the new method.